+++ Mlem/Navigation/AnyNavigationPath.swift
@@ -12,6 +12,7 @@ protocol AnyNavigationPath {
     
     associatedtype Route: Routable
     
+    /// Implementation should make a route that makes sense for the passed-in data value and can be appended to the navigation path.
     static func makeRoute<V>(_ value: V) -> Route where V: Hashable
     
     /// The number of elements in this path.
